+++ assets/script/verse.js
@@ -3,10 +3,8 @@ const verses = [...document.getElementsByClassName("current-verse")]
 window.onhashchange = () =>
   verses.map((verse) => {
     const curr = window.location.hash.substring(1)
-    if (curr.length > 0)
-      verse.innerText = ":" + curr
-      else
-      verse.innerText = ""
+    if (curr.length > 0) verse.innerText = ":" + curr
+    else verse.innerText = ""
   })
 window.onhashchange()